navbar refactoring
